refactor: fold overview view code into monitor/wandb.py
